Downloads/Smart-parking_management_App-main (2).zip done/Smart-parking_management_App-main/backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
mysql-connector-python>=8.0
Werkzeug>=2.0
python-dotenv>=0.21.0
gunicorn>=21.2.0
